Keyhunter Dean-Pierce
The code is taken from the Dean Pierce repository https://github.com/pierce403/keyhunter, thank him for this! And adapted to fit your needs. The code is kept for myself, but is available to those who are looking for a solution to run it.
Added: 
Enter search directory
Processing file indication
Output the result to a text file